requirements.txt
python-dotenv==1.0.1
streamlit==1.37.1
weaviate-client==4.10.4
plotly